+++ web/routes/index.tsx
@@ -9,6 +9,8 @@ import CartPage from "../pages/Cart";
 import CheckoutPage from "../pages/Checkout";
 import OrderList from "../pages/OrderList";
 import OrderDetail from "../pages/OrderDetail";
+import Payment from "../pages/Payment";
+import PaymentResult from "../pages/PaymentResult";
 
 export default function AppRoutes() {
   return (
@@ -22,6 +24,8 @@ export default function AppRoutes() {
       <Route path="/checkout" element={<CheckoutPage />} />
       <Route path="/orders" element={<OrderList />} />
       <Route path="/orders/:id" element={<OrderDetail />} />
+      <Route path="/orders/:orderId/pay" element={<Payment />} />
+      <Route path="/payment/result" element={<PaymentResult />} />
       <Route path="/profile" element={<Profile />} />
       <Route path="*" element={<Navigate to="/" replace />} />
     </Routes>